+++ src/lsp.ts
@@ -29,8 +29,6 @@ import {
   CLIENT_ID,
   CLIENT_NAME,
   DIAGNOSTIC_COLLECTION_NAME,
-  MIN_VERSION,
-  LATEST_VERSION,
 } from "./constants";
 import { Environment } from "./env";
 import { rulesRefreshed } from "./lspExtensions";
@@ -40,6 +38,7 @@ import {
   ProxyOutputChannel,
   withSentryAsync,
 } from "./telemetry/sentry";
+import { checkCliVersion } from "./utils";
 
 async function findSemgrep(env: Environment): Promise<Executable | null> {
   let server_path = which.sync(SEMGREP_BINARY, { nothrow: true });
@@ -126,21 +125,8 @@ async function serverOptionsCli(
   if (!env.config.cfg.get("ignoreCliVersion")) {
     const cmd = `"${server.command}" --version`;
     const version = await execShell(cmd, server.options?.env);
-    const minor = semver.minor(version);
-    const major = semver.major(version);
-    vscode.commands.executeCommand("setContext", "semgrep.cli.minor", minor);
-    vscode.commands.executeCommand("setContext", "semgrep.cli.major", major);
-    if (!semver.satisfies(version, MIN_VERSION)) {
-      vscode.window.showErrorMessage(
-        `The Semgrep Extension requires a Semgrep CLI version ${MIN_VERSION}, the current installed version is ${version}, please upgrade.`,
-      );
-      return null;
-    }
-    if (!semver.satisfies(version, LATEST_VERSION)) {
-      vscode.window.showWarningMessage(
-        `Some features of the Semgrep Extension require a Semgrep CLI version ${LATEST_VERSION}, but the current installed version is ${version}, some features may be disabled, please upgrade.`,
-      );
-    }
+    const semVersion = new semver.SemVer(version);
+    checkCliVersion(semVersion);
     env.semgrepVersion = version;
     await env.reloadConfig();
   }